requirements.txt
Flask==2.1.2
requests==2.32.3
pycryptodome==3.10.1
logging
subprocess
socket
base64
time
